+++ Modules/Intent.Modules.Common.CSharp/Builder/ICSharpMethodDeclarationActual.cs
@@ -20,4 +20,5 @@ public interface ICSharpMethodDeclaration<out TCSharpMethodDeclaration> : ICShar
     TCSharpMethodDeclaration New();
     TCSharpMethodDeclaration Virtual();
     TCSharpMethodDeclaration Abstract();
+    TCSharpMethodDeclaration WithoutMethodModifier();
 }